+++ src/Classifiers/LogisticRegression.php
@@ -63,7 +63,7 @@ class LogisticRegression implements Estimator, Learner, Online, Probabilistic, V
     protected $optimizer;
 
     /**
-     * The strength of the L2 regularization penalty.
+     * The amount of L2 regularization applied to the weights of the output layer.
      *
      * @var float
      */