+++ src/Classifiers/MultilayerPerceptron.php
@@ -81,7 +81,7 @@ class MultilayerPerceptron implements Estimator, Learner, Online, Probabilistic,
     protected $optimizer;
 
     /**
-     * The strength of the L2 regularization penalty.
+     * The amount of L2 regularization applied to the weights of the output layer.
      *
      * @var float
      */