+++ src/NeuralNet/Layers/Multiclass.php
@@ -40,7 +40,7 @@ class Multiclass implements Output, Parametric
     ];
 
     /**
-     * The L2 regularization amount.
+     * The amount of L2 regularization applied to the weights.
      *
      * @var float
      */
@@ -89,21 +89,21 @@ class Multiclass implements Output, Parametric
     protected $biases;
 
     /**
-     * The memoized input matrix.
+     * The memorized input matrix.
      *
      * @var \Tensor\Matrix|null
      */
     protected $input;
 
     /**
-     * The memoized z matrix.
+     * The memorized z matrix.
      *
      * @var \Tensor\Matrix|null
      */
     protected $z;
 
     /**
-     * The memoized activation matrix.
+     * The memorized activation matrix.
      *
      * @var \Tensor\Matrix|null
      */
@@ -119,7 +119,7 @@ class Multiclass implements Output, Parametric
      */
     public function __construct(
         array $classes,
-        float $alpha = 1e-4,
+        float $alpha = 0.0,
         ?ClassificationLoss $costFn = null,
         ?Initializer $weightInitializer = null,
         ?Initializer $biasInitializer = null
@@ -261,12 +261,14 @@ class Multiclass implements Output, Parametric
                 ->multiply($dL);
         }
 
+        $dW = $dA->matmul($this->input->transpose());
+        $dB = $dA->sum();
+
         $weights = $this->weights->param();
 
-        $dW = $dA->matmul($this->input->transpose())
-            ->add($weights->multiply($this->alpha));
-
-        $dB = $dA->sum();
+        if ($this->alpha) {
+            $dW = $dW->add($weights->multiply($this->alpha));
+        }
 
         $this->weights->update($optimizer->step($this->weights, $dW));
         $this->biases->update($optimizer->step($this->biases, $dB));