+++ src/Regressors/Adaline.php
@@ -63,7 +63,7 @@ class Adaline implements Estimator, Learner, Online, Verbose, Persistable
     protected $optimizer;
 
     /**
-     * The strength of the L2 regularization penalty.
+     * The amount of L2 regularization applied to the weights of the output layer.
      *
      * @var float
      */