+++ src/Regressors/MLPRegressor.php
@@ -79,7 +79,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
     protected $optimizer;
 
     /**
-     * The strength of the L2 regularization penalty.
+     * The amount of L2 regularization applied to the weights of the output layer.
      *
      * @var float
      */